add: index.html
It contains a simple UI that has a form to input a link and a button to
submit the link. There is a function handleSubmit() that is called when
the submit button is clicked and there is a sample snippet on how to
call an endpoint there, but it doesn't really do anything.